Update trivy.yml
I've removed the working-directory line from the Build an image from Dockerfile step. This will tell the runner to look for the Dockerfile in the repository's root, which should resolve the error. If the Dockerfile is in a different subdirectory, you will need to add the correct path back in.