Update trivy.yml
The main change was updating the uses line for the trivy-action step. Using a stable version tag like v0.32.0 is a best practice that should prevent this kind of error in the future. I'll keep the workflow open for any other changes you'd like to make. Let me know if you want to add more checks or change the schedule!